% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer_type.R
\name{answer_type}
\alias{answer_type}
\title{Answer Type Codes and Labels for HMIS Field Responses}
\usage{
answer_type(answer_code, ..., full_label = FALSE, .FY = 24)
}
\arguments{
\item{answer_code}{\itemize{
\item The "Answer Type" as a string;
\itemize{
\item This includes \code{"unknown"} which will return the unknown answer types: \code{c(8, 9, 99)} \emph{("Client doesn’t know", "Client prefers not to answer", and "Data not collected".)}.
}
\item The HMIS CSV Code as an integer; or
\item \code{"ls"} to get the full list of Answer Types.
}}

\item{...}{These dots are for future extensions and must be empty.}

\item{full_label}{If set to \code{TRUE}, \code{answerType()} will return the full text
of the answer type rather than the shorthand label (i.e., \code{"Client prefers not to answer"} instead of \code{"noAnswer"}). \strong{Defaults to \code{FALSE}.}}

\item{.FY}{The HMIS Data Standards Fiscal Year (entered numerically with two
digits, i.e., \code{22} for FY22). \emph{\strong{Defaults to FY24.}} Can be set backwards
up to FY22 for backwards compatibility with older HMIS Extracts.}
}
\value{
Depending on the input: The "Answer Type" label; The HMIS CSV Code;
or the full list of Answer Types.
}
\description{
If the input is an Answer Type, returns the HMIS CSV Code for that Answer
Type. If the input is an HMIS CSV Code, returns the name of the Answer Type.
If the input is \code{"ls"} returns the full list of Answer Type labels.
}
