% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_inventory_file.R
\name{append_inventory_file}
\alias{append_inventory_file}
\title{Append the Inventory.csv File to the HMIS Extract List}
\usage{
append_inventory_file(.data)
}
\arguments{
\item{.data}{The HMIS extract list.}
}
\value{
As a list, the HMIS Extract with the Inventory.csv file included.
}
\description{
This function reads the \code{Inventory.csv} file and \code{\link[base:append]{base::append()}}s it to the
supplied HMIS Extract.
}
