% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_chronicity.R
\name{calculate_chronicity_at_entry}
\alias{calculate_chronicity_at_entry}
\title{Add a \code{Chronic_Homelessness_Status_At_Entry} Column to a Data Frame}
\usage{
calculate_chronicity_at_entry(.data)
}
\arguments{
\item{.data}{A tibble or data frame that is, corresponds to, or has all the
data from the \code{client} tibble in the \verb{HMIS Data Export}.}
}
\value{
the tibble used as the function input, with a new column called
\code{Chronic_Homelessness_Status_At_Entry}.
}
\description{
Adds a .csv \code{Chronic_Homelessness_Status_At_Entry} column to a tibble or data
frame using \code{\link[dplyr:mutate]{dplyr::mutate()}}.
}
