% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_spm_details.R
\name{calculate_spm_details}
\alias{calculate_spm_details}
\title{Calculate HUD System Performance Measures}
\usage{
calculate_spm_details(hmis_extract = NULL, spm, ..., fiscal_year)
}
\arguments{
\item{hmis_extract}{The HMIS Extract to use (may be piped in, or if left
null, this function will call \link{load_hmis("newest")}).}

\item{spm}{As a character (in "quotes" if there is a letter, i.e., "2a"),
which SPM (1-7b). You may select more than one, i.e., \code{c(1, "2b", 5, "7a")}}

\item{...}{These dots are for future extensions and must be empty.}

\item{fiscal_year}{The federal fiscal year for which to run the SPMs.}
}
\value{
One or more SPMs.
}
\description{
Use an HMIS data export to calculate one or more HUD System Performance
Measures for a given federal fiscal year.
}
