% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pit.R
\name{check_pit}
\alias{check_pit}
\title{Check Enrollments on a Specific Day}
\usage{
check_pit(
  hmis_extract,
  pit,
  ...,
  project_type = NULL,
  project_id = NULL,
  dedup_only = FALSE
)
}
\arguments{
\item{hmis_extract}{the full HMIS Extract List.}

\item{pit}{the date for the PIT count.}

\item{...}{These dots are for future extensions and must be empty.}

\item{project_type}{OPTIONAL: runs a PIT count only on the specified project
type.}

\item{project_id}{OPTIONAL: runs a PIT count only on the specified project
per its \code{ProjectID} number.}

\item{dedup_only}{OPTIONAL: if set to \code{TRUE} returns only the deduplicated
data frame (\code{tibble}) instead of the list of the total PIT count and the
deduplicated version. Defaults to \code{FALSE}.}
}
\value{
a list of 2 \code{tibble}s: entryExitData & deduplicatedPeople (or, if
\code{dedup_only = TRUE}, then just the deduplicated PIT count tibble).
}
\description{
Returns two tibbles showing total enrollments and deduplicated client PIT
counts on a specific day per any optional arguments.
}
