% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplicate.R
\name{deduplicate}
\alias{deduplicate}
\title{Deduplicate HMIS Data to One Row Per Unique SPID}
\usage{
deduplicate(.data, ..., arrange_by = "EntryDate")
}
\arguments{
\item{.data}{The tibble / data frame to be deduplicated.}

\item{...}{These dots are for future extensions and must be empty.}

\item{arrange_by}{The value to arrange in descending order before the
distinct function from the dplyr package is applied to deduplicate the
data. This defaults to EntryDate.}
}
\value{
The deduplicated tibble / data frame that was originally given to the
function.
}
\description{
This function returns the data frame input into it with only one row for each
unique PersonalID. It defaults to sorting the data descending by EntryDate
for this purpose, if not otherwise specified.
}
