% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_table.R
\name{easy_table}
\alias{easy_table}
\title{\code{easy_table()} Gives You a "Tibble" Formatted Table for a Column of Your Data}
\usage{
easy_table(.data, .col, ..., na_last = TRUE, ignore_na = FALSE)
}
\arguments{
\item{.data}{your data}

\item{.col}{the column for which you want to \emph{see} a table.}

\item{...}{These dots are for future extensions and must be empty.}

\item{na_last}{Defaults to \code{TRUE} and puts \code{NA} values last in the table
regardless of the value of their total \code{Count}. If set to \code{FALSE}, \code{NA}
will appear in the table where it falls according to its \code{Count} value.}

\item{ignore_na}{Defaults to \code{FALSE} and sets the \code{useNA} argument in
\code{\link[=table]{table()}} to \code{"ifany"}. If set to \code{TRUE} \code{\link[=table]{table()}} will run without the
\code{useNA} argument.}
}
\value{
Returns a table of \code{.col} from \code{.data} with the \code{useNA = "ifany"}
argument used unless \code{ignore_na} is set to \code{TRUE}.
}
\description{
Returns a \code{table()} of a column of your data frame, formatted as a \code{tibble},
pre-sorted descending by count with \code{NA} last unless otherwise specified.
}
\examples{
\dontrun{

hmis <- load_hmis("newest")

# This will return a tibble of `Race_Ethnicity` in the Client file
# with the count of any NA's last
hmis$client \%>\% easy_table(Race_Ethnicity)

# This will return a tibble of `Race_Ethnicity` in the Client file
# with the count of any NA's in descending order with other categories
hmis$client \%>\% easy_table(Race_Ethnicity, na_last = FALSE)

# This will return a tibble of `Race_Ethnicity` in the Client file
# excluding the count of any NA's
hmis$client \%>\% easy_table(Race_Ethnicity, ignore_na = TRUE)

}
}
