% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECHO-package.R
\docType{package}
\name{ECHO-package}
\alias{ECHO}
\alias{ECHO-package}
\title{ECHO: Internal ECHO Tools for Analyzing HMIS Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package contains reusable functions for common repetitive tasks when reading, viewing, cleaning, and analyzing HMIS data, especially from the .csv files in the HMIS XML Exports.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.austinecho.org}
  \item \url{https://github.com/echoatx/echo-rpkg}
  \item Report bugs at \url{https://github.com/echoatx/echo-rpkg/issues}
}

}
\keyword{internal}
