% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo_help.R
\name{echo_help}
\alias{echo_help}
\title{Open the ECHO Package Help Website}
\usage{
echo_help(function_name = NULL)
}
\value{
Depending on the input: The "Answer Type" label; The HMIS CSV Code;
or the full list of Answer Types.
}
\description{
This function opens the help website for the ECHO Package.
}
