% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entry_situation.R
\name{entry_situation}
\alias{entry_situation}
\title{Entry Situation HMIS Codes, Labels and Data Frame}
\usage{
entry_situation(input, flag = NULL, ..., .FY = 24)
}
\arguments{
\item{input}{The "Entry Situation Label" as a string; The HMIS CSV Code as an
integer; "ls" to get the full list of Entry Situation Labels; or "load" to
get the full Entry Situation data frame.}

\item{flag}{OPTIONAL: For the FY24 and later versions of this function,
adding the flag \"category\" will return the category of the input value
(i.e. homeless situation, permanent housing destination, etc.)}

\item{...}{These dots are for future extensions and must be empty.}

\item{.FY}{The HMIS data standards fiscal year (defaults to current).}
}
\value{
Depending on the input: The "Entry Situation" label; The HMIS CSV
Code; the list of Entry Situation labels; or the Entry Situation Labels
data frame.
}
\description{
If the input is an Entry Situation, returns the HMIS CSV Code for that entry
situation. If the input is an HMIS CSV Code, returns the Entry Situation
Label. If the input is "ls" returns the list of Entry Situation labels. If
the input is "load" returns the full Entry Situations data frame.
}
