% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix_pit.R
\name{\%pit\%}
\alias{\%pit\%}
\title{The "PIT Pipe" is a Shortcut to Get Enrollments on a Specific Day from an
Existing Dataframe/Tibble}
\usage{
.data \%pit\% .pit
}
\arguments{
\item{.data}{your data}

\item{.pit}{the PIT date}
}
\value{
The deduplicated (\link{\%pit\%}) or full (\link{\%pit-enrollments\%}) PIT data
frame for the piped data.
}
\description{
Returns two tibbles showing total enrollments and deduplicated client PIT
counts on a specific day per any optional arguments.
}
