% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix_see.R
\name{\%see\%}
\alias{\%see\%}
\title{The "See Pipe" Lets You \emph{See} a \code{\link[base:table]{base::table()}} of a Column in a Data Frame}
\usage{
.data \%see\% .col
}
\arguments{
\item{.data}{your data}

\item{.col}{the column for which you want to \emph{see} a table.}
}
\value{
Returns a table of \code{.col} from \code{.data} with the \code{useNA = "ifany"}
argument.
}
\description{
Returns a table with the \code{useNA = "ifany"} argument showing a column of your
data frame.
}
