% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix_unlock.R
\name{\%unlock\%}
\alias{\%unlock\%}
\title{Decrypt Confidential Columns}
\usage{
lhs \%unlock\% rhs
}
\arguments{
\item{lhs}{the tibble with encrypted columns.}

\item{rhs}{the column or columns to decrypt. Put the Column names in quotes
(""). If decrypting multiple columns, put all quoted colmn names in \code{\link[=c]{c()}}.
You will need to enter the password for each column you are decrypting.}
}
\value{
returns the original tibble with the encrypted columns unlocked.
}
\description{
Decrypt one or more columns in a \code{tibble}. Note: you must enter the password
once for each column you are decrypting.
}
