% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_overview_functions.R
\name{hrs_clients_housed}
\alias{hrs_clients_housed}
\title{HRS Clients Housed}
\usage{
hrs_clients_housed(.data, reporting_period_start, reporting_period_end)
}
\arguments{
\item{.data}{the data frame to which to apply this function.}

\item{reporting_period_start}{the start of the reporting period.}

\item{reporting_period_end}{the end of the reporting period.}
}
\value{
two tibbles: \code{clients_housed_grouped} and \code{annual_housings}.
}
\description{
This function finds clients housed by project type and for the previous 12
months for the specified time period.
}
