% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_overview_functions.R
\name{hrs_enrollments}
\alias{hrs_enrollments}
\title{HRS Enrollments Table Generator}
\usage{
hrs_enrollments(
  hmis_extract = NULL,
  ...,
  reporting_period_start,
  reporting_period_end,
  extractdate = NULL
)
}
\arguments{
\item{hmis_extract}{the HMIS CSV Extract.}

\item{...}{These dots are for future extensions and must be empty.}

\item{reporting_period_start}{The start date of the reporting period.}

\item{reporting_period_end}{The end date of the reporting period.}

\item{extract_date}{The date of the HMIS CSV Extract. (Defaults to the
extract date of \code{hmis_extract}, but it can be set manually)}
}
\value{
A tibble with an estimate of all of the sheltered clients as of the
\code{report_effective_date}.
}
\description{
This function returns a tibble with all of the unsheltered clients from the
CA Reference list that have not touched the system in the past 180 days, per
the information available in the CA Reference List.
}
