% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_overview_functions.R
\name{hrs_returns_demographics}
\alias{hrs_returns_demographics}
\title{HRS Returns Demographics}
\usage{
hrs_returns_demographics(
  hmis_extract = NULL,
  ...,
  effective_date,
  max_days_to_reappear = 730,
  return_function_info = FALSE,
  return_orig_return_window = FALSE
)
}
\arguments{
\item{hmis_extract}{the HMIS CSV Extract.}

\item{...}{These dots are for future extensions and must be empty.}

\item{effective_date}{the effective date for the returns report (usually the
dashboard date). (The interval will be set for \code{effective_date - 730} days
through \code{effective_date - 1} day.)}

\item{max_days_to_reappear}{The return window. Defaults to \code{730} days (two
years).}

\item{return_function_info}{If set to \code{TRUE}, the function will return a
status message about the internal variable (dates) it is using (instead of
the data frame it normally returns). Defaults to \code{FALSE.}}

\item{return_orig_return_window}{If set to \code{TRUE}, the function will return
the tibble containing the original exits being used to calculate returns,
instead of the resulting returns table. Defaults to \code{FALSE.}}
}
\value{
a tibble: \code{returns_report}.
}
\description{
This function finds the demographics (Race/Ethnicity) and Project Types for
people who returned to homelessness within 2 years (default - edit
\code{max_days_to_reappear} to override this).
}
