% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_overview_functions.R
\name{hrs_unsheltered_snapshot}
\alias{hrs_unsheltered_snapshot}
\title{HRS Unsheltered Snapshot}
\usage{
hrs_unsheltered_snapshot(
  hmis_extract = NULL,
  ...,
  dashboard_date,
  ce_bnl,
  bnl_date,
  mean_days = 180
)
}
\arguments{
\item{hmis_extract}{the HMIS CSV Extract.}

\item{...}{These dots are for future extensions and must be empty.}

\item{dashboard_date}{The date for which the dashboard is being generated.}

\item{ce_bnl}{the file path to the CE BNL Excel file.}

\item{bnl_date}{The date of the CE BNL (do not supply an actual date value,
just a \code{string}).}

\item{mean_days}{The lookback timeframe. Defaults to \code{180} days (six months).}
}
\value{
A tibble with all of the unsheltered clients from the CA Reference
list that have not touched the system in the past \code{mean_days}.
}
\description{
This function returns a tibble with all of the unsheltered clients from the
CA Reference list who have not touched the system in the past 180 days, per
the information available in the CA Reference List.
}
