% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HUD_CoC_Program_Specific_Data_Elements_Table_FY24}
\alias{HUD_CoC_Program_Specific_Data_Elements_Table_FY24}
\title{CoC Program Specific Data Elements Table (FY24)}
\format{
\subsection{\code{HUD_CoC_Program_Specific_Data_Elements_Table_FY24} A data frame with 7 rows and 19 columns:}{

\describe{
\item{CoC Intervention}{String: The Project Type (and sub-category, if applicable).}
\item{Income and Sources}{Boolean: is this element required for the CoC Intervention in this row?}
\item{Non-Cash Benefits}{Boolean: is this element required for the CoC Intervention in this row?}
\item{Health Insurance}{Boolean: is this element required for the CoC Intervention in this row?}
\item{Physical Disability}{Boolean: is this element required for the CoC Intervention in this row?}
\item{Developmental Disability}{Boolean: is this element required for the CoC Intervention in this row?}
\item{Chronic Health Condition}{Boolean: is this element required for the CoC Intervention in this row?}
\item{HIV/AIDS}{Boolean: is this element required for the CoC Intervention in this row?}
\item{Mental Health Disorder}{Boolean: is this element required for the CoC Intervention in this row?}
\item{Substance Use Disorder}{Boolean: is this element required for the CoC Intervention in this row?}
\item{Domestic Violence}{Boolean: is this element required for the CoC Intervention in this row?}
\item{Current Living Situation}{Boolean: is this element required for the CoC Intervention in this row?}
\item{Date of Engagement}{Boolean: is this element required for the CoC Intervention in this row?}
\item{Coordinated Entry Assessment}{Boolean: is this element required for the CoC Intervention in this row?}
\item{Coordinated Entry Event}{Boolean: is this element required for the CoC Intervention in this row?}
\item{Moving On Assistance Provided}{Boolean: is this element required for the CoC Intervention in this row?}
\item{Translation Assistance Needed}{Boolean: is this element required for the CoC Intervention in this row?}
\item{Sexual Orientation}{Boolean: is this element required for the CoC Intervention in this row?}
\item{Housing Assessment at Exit}{Boolean: is this element required for the CoC Intervention in this row?}
...
}
}
}
\source{
\url{https://files.hudexchange.info/resources/documents/CoC-Program-HMIS-Manual-2024.pdf}
}
\usage{
HUD_CoC_Program_Specific_Data_Elements_Table_FY24
}
\description{
This table identifies which PSDEs are required for each CoC Component ...
}
\keyword{datasets}
