% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HUD_HMIS_Universal_Data_Elements_Table_FY24}
\alias{HUD_HMIS_Universal_Data_Elements_Table_FY24}
\title{FY24 HUD HMIS Universal Data Elements Table}
\format{
\subsection{\code{HUD_HMIS_Universal_Data_Elements_Table_FY24} A data frame with 27 rows and 3 columns:}{

\describe{
\item{UDE}{The Name of the UDE}
\item{HMIS Extract}{Where in the HMIS Extract does this UDE appear as a column?}
\item{CSV File}{In which .csv file in the HMIS data export does this UDE appear as a column?}
...
}
}
}
\source{
\url{https://files.hudexchange.info/resources/documents/HMIS-Data-Dictionary-2024.pdf}
}
\usage{
HUD_HMIS_Universal_Data_Elements_Table_FY24
}
\description{
This table lists the items that are considered UDEs and identifies where they can be found in the HMIS Extract and/or the
HMIS data export .csv files ...
}
\keyword{datasets}
