% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HUD_LivingSituations_Destinations_SubsidyTypes_FY24}
\alias{HUD_LivingSituations_Destinations_SubsidyTypes_FY24}
\title{FY24 HUD HMIS Living Situation & Destination Option List with Rental Subsidy Types}
\format{
\subsection{\code{HUD_LivingSituations_Destinations_SubsidyTypes_FY24} A data frame with 43 rows and 3 columns:}{

\describe{
\item{Value}{HMIS CSV code}
\item{Description}{Text explanation of the CSV code}
\item{Classification}{HUD's classification of the value's category (permanent housing, homelessness, etc.)}
...
}
}
}
\source{
\url{https://files.hudexchange.info/resources/documents/HMIS-CSV-Format-Specifications-2024.pdf}
}
\usage{
HUD_LivingSituations_Destinations_SubsidyTypes_FY24
}
\description{
This provides the CSV codes and text labels for the living situations/ exit destinations and rental subsidy types for the
FY24 HMIS data standards ...
}
\keyword{datasets}
