% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_bnl.R
\name{import_bnl}
\alias{import_bnl}
\title{Imports a CE BNL into R from its downloaded Excel file format.}
\usage{
import_bnl(downloaded_bnl, list_date, ..., autoclean = TRUE, as.dedup = FALSE)
}
\arguments{
\item{downloaded_bnl}{The \code{.xls }or \code{.xlsx} BNL file -or- the folder with the
individual \code{.csv} files of the BNL.}

\item{list_date}{The date of the BNL.}

\item{...}{These dots are for future extensions and must be empty.}

\item{autoclean}{Automatically filters down the full BNL to include only
certain columns and cleans data if set to \link{TRUE} (this is the default). If
set to \code{FALSE}, the BNL will be imported verbatim from Excel into R.}

\item{as.dedup}{Automatically deduplicates the BNL if set to \code{TRUE}. Defaults
to \code{FALSE}. This keeps one instance of each PersonalID based on the most
recent CADate ("API Assessment Date," if available and "ID Date" if not).}
}
\value{
Returns a list of the CE BNL data frames in tibble format with the
BNL Date.
}
\description{
Imports a CE BNL into R from its downloaded Excel file format.
}
