% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_hmis.R
\name{import_hmis}
\alias{import_hmis}
\title{Read in an HMIS CSV/XML Export}
\usage{
import_hmis(
  extract_path,
  ...,
  include_disabilities = FALSE,
  all_files = FALSE,
  suppressNormalReadrCsvWarnings = TRUE,
  runningStandalone = TRUE,
  .FY = 24
)
}
\arguments{
\item{extract_path}{the path to the folder containing the HMIS CSV/XML Export
.csv files.}

\item{...}{These dots are for future extensions and must be empty.}

\item{include_disabilities}{OPTIONAL: specifies whether to import the
disabilities file (defaults to \code{FALSE}).}

\item{.FY}{The HMIS Data Standards Fiscal Year (entered numerically with two
digits, i.e., \code{22} for FY22). \emph{\strong{Defaults to FY24.}} Can be set backwards
up to FY22 for backwards compatibility with older HMIS Extracts.}

\item{extract_date}{the date of the HMIS CSV/XML Export.}
}
\value{
a list of 8-9 tibbles: client, services, entry, exit, project,
healthanddv, incomebenefits, & organization (and disabilities if
include_disabilities = TRUE).
}
\description{
Reads in .csv files from data exported from HMIS per the specified date and
returns a list of tibbles.
}
