% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_echo_themes.R
\name{install_echo_themes}
\alias{install_echo_themes}
\title{Install or Update the Custom ECHO RStudio Themes}
\usage{
install_echo_themes(theme)
}
\arguments{
\item{theme}{Type \code{"light"} or \code{"dark"} to install the light theme or the
dark theme. You can install both, but they must be installed separately so
you'll have to run this function once for each.}
}
\value{
Installs or updates the selected custom ECHO RStudio theme (\code{"light"}
or \code{"dark"}).
}
\description{
This function installs/updates the custom ECHO themes for the editor in
RStudio
}
