% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_vsp.R
\name{is_vsp}
\alias{is_vsp}
\title{Do Any Supplied Project IDs Belong to a VSP?}
\usage{
is_vsp(
  project_id,
  ...,
  alternate_database_users_only = FALSE,
  hmis_extract = NULL
)
}
\arguments{
\item{project_id}{A single Project ID or multiple Project IDs (this function
is vectorized).}

\item{...}{These dots are for future extensions and must be empty.}

\item{alternate_database_users_only}{Defaults to \code{FALSE}. If set to \code{TRUE}
this will only return \code{TRUE} for VSPs that use an alternate database
instead of HMIS.}

\item{hmis_extract}{The full CSV Extract List. This will Automatically use
"hmis" if it is in the Global Environment, but otherwise you must specify
the hmis_extract variable.}
}
\value{
\code{TRUE} or \code{FALSE} for each supplied Project ID, depending on whether
it belongs to a VSP organization. (If \code{alternate_database_users_only} is
set to \code{TRUE}, then any VSPs that use HMIS and not an alternate database
instead will aslo return \code{FALSE}.)
}
\description{
Determines whether the Project IDs belong to a VSP, and optionally whether
that VSP uses an alternate database instead of HMIS.
}
