% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_extract_files.R
\name{join_extract_files}
\alias{join_extract_files}
\title{Quickly Join The Entry/Exit/Client/Project Files}
\usage{
join_extract_files(.data, ..., reporting_period = NULL)
}
\arguments{
\item{.data}{the HMIS CSV/XML extract list of tibbles.}

\item{...}{These dots are for future extensions and must be empty.}

\item{reporting_period}{If set to a \code{\link[lubridate:interval]{lubridate::interval()}},
\code{join_extract_files()} will use \code{\link[dplyr:filter]{dplyr::filter()}} to filter for
\code{EntryDate}s within the given interval.}
}
\value{
As a \code{tibble}, the joined Entry/Exit/Client/Project files.
}
\description{
This will use \code{\link[dplyr:mutate-joins]{dplyr::left_join()}} to join the Entry, Exit, Client, and
Project files of the HMIS Extract, optionally filtering for entry dates
within a specific reporting period.
}
