% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_bnl.R
\name{load_bnl}
\alias{load_bnl}
\title{Loads a BNL from .rda file format into the Environment.}
\usage{
load_bnl(which_bnl, ..., as.full = FALSE, as.dedup = FALSE)
}
\arguments{
\item{which_bnl}{the path to the BNL \code{.rda} file you want to load, or the
date of the BNL you want to load (\code{"YYYY-MM-DD"}), -or- type \code{"newest"} to
load the most recent BNL.}

\item{...}{These dots are for future extensions and must be empty.}

\item{as.full}{If using the \code{"newest"}, parameter, set \code{as.full} to \code{TRUE} if
you need to load the full BNL, not the autocleaned version. This defaults
to \code{FALSE} and loads the autocleaned version.}

\item{as.dedup}{automatically deduplicates the BNL if set to \code{TRUE}. Defaults
to \code{FALSE}. This keeps one instance of each PersonalID based on the most
recent CADate ("API Assessment Date," if available and "ID Date" if not).}
}
\value{
loads the selected CE BNL \code{.rda} file.
}
\description{
Returns the list of tibbles saved in the selected BNL ".rda" file.
}
