% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_race_details.R
\name{load_race_details}
\alias{load_race_details}
\alias{attach_race_details}
\title{Loads or Attaches Client Races/Ethnicity by PersonalID from the HMIS CSV/XML
Export.}
\usage{
load_race_details(hmis_extract = NULL, hmis_extract_full = NULL)

attach_race_details(.data, hmis_extract = NULL, hmis_extract_full = NULL)
}
\arguments{
\item{hmis_extract}{the hmis extract to use for these functions. If none is
supplied, the function will try to find an HMIS extract named \code{hmis} in
your environment. If it cannot do so, it will then try to find any HMIS
Extract in your environment to use (if there is more than one it will
select the first one it finds). If it cannot find an extract, it will call
\link{load_hmis("newest")} to use that, and it will inform you that it did so.}

\item{.data}{for \code{attach_race_details()}, the tibble or data frame you want
the full race/ethnicity data attached to.}
}
\value{
\code{\link[=load_race_details]{load_race_details()}} returns a data frame with the Race/Ethnicity
answers for each client in the \code{Client.csv} file associated with the HMIS
Extract used for the \code{hmis_data} argument. \code{\link[=attach_race_details]{attach_race_details()}} calls
\code{\link[=load_race_details]{load_race_details()}} and then uses the Tidyverse's \code{\link[dplyr:mutate-joins]{dplyr::left_join()}}
function to add the Race/Ethnicity data for each client in \code{.data} from
their answers in the \code{Client.csv} file associated with the HMIS Extract
used for the \code{hmis_data} argument.
}
\description{
\code{\link[=load_race_details]{load_race_details()}} returns a data frame containing all race answers per
PersonalID in the \code{Client.csv} file associated with the HMIS Extract you are
using.
}
\details{
\code{\link[=attach_race_details]{attach_race_details()}} attaches those details to a data frame, using
Tidyverse's \code{\link[dplyr:mutate-joins]{dplyr::left_join()}} function (by \code{PersonalID}) a data frame
containing all race answers per \code{PersonalID} in the \code{Client.csv} file
associated with the HMIS Extract you are using.
}
