% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_universe.R
\name{make_universe}
\alias{make_universe}
\title{Generate the Enrollment and/or Client Universe(s)}
\usage{
make_universe(
  hmis_extract = NULL,
  ...,
  universe_type,
  timeframe,
  calendar_year_start = NULL,
  old_version = FALSE,
  .FY = 24
)
}
\arguments{
\item{hmis_extract}{the HMIS CSV/XML extract list of tibbles.}

\item{...}{These dots are for future extensions and must be empty.}

\item{universe_type}{which type of universe: \code{"enrollment"}, \code{"client"}, or
\code{"both"}. (If both, you can use dollar sign notation to get which universe
you want from the object created: i.e., \code{your_universe$enrollment} /
\code{your_universe$client}.)}

\item{timeframe}{a lubridate interval: the date the lookback period begins
through the date the lookback period ends. (If you enter \code{"ytd"} instead,
the function will use January first of the current year through the last
day of the previous full month.)}

\item{calendar_year_start}{the calendar year for the analysis: this defaults
to \code{calendarYearStart} if it is in your environment, or January first of
the year of the start of the reporting period if \code{calendarYearStart} is not
in your environment. Otherwise, you can specify it as a string:
\code{"YYYY-01-01"}.}

\item{old_version}{Defaults to \code{FALSE.} If set to \code{TRUE} it will use the
original code which does not auto filter out duplicate projects in the
projects file.}

\item{.FY}{The HMIS Data Standards Fiscal Year (entered numerically with two
digits, i.e., \code{22} for FY22). \emph{\strong{Defaults to FY24.}} Can be set backwards
up to FY22 for backwards compatibility with older HMIS Extracts.}
}
\value{
A tibble: the enrollment universe.
}
\description{
Create a tibble of enrollments that overlap with the reporting period.
}
