% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybeChronic_list.R
\name{maybeChronic_list}
\alias{maybeChronic_list}
\title{Chronicity Calculator for a Data Frame of Clients}
\usage{
maybeChronic_list(entry_file, extract_date)
}
\arguments{
\item{entry_file}{the 'entry' data frame from the list generated/returned by
the import_hmis function.}

\item{extract_date}{the date the CSV/XML extract was pulled from HMIS:
YYYY-MM-DD.}
}
\value{
List of  \code{TRUE} or \code{FALSE} for each row of the Entry.csv file for
whether that row is estimated to be Chronic.
}
\description{
Estimates whether a client may be experiencing chronic homelessness per the
available information in the HMIS export.
}
