% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybeChronic_spid.R
\name{maybeChronic_spid}
\alias{maybeChronic_spid}
\title{Chronicity Calculator for an Individual Client}
\usage{
maybeChronic_spid(hmis_extract = NULL, spid, extract_date = NULL)
}
\arguments{
\item{hmis_extract}{The full \verb{CSV Extract} List. This will Automatically use
\code{hmis} if it is in the Global Environment. It can also be piped in.
Otherwise, it will search the global environment and use the first HMIS
extract it finds, or it will call \link[=load_hmis]{load_hmis("newest")} and
use that if there is no HMIS extract in the environment to use.}

\item{spid}{the client's SPID: the value in the \code{PersonalID} column.}

\item{extractdate}{the date the CSV/XML extract was pulled from HMIS:
YYYY-MM-DD. defaults to \code{hmis_extract$extractDate}.}
}
\value{
TRUE or FALSE.
}
\description{
Estimates whether a single client may be experiencing chronic homelessness
per the available information in their most recent entry in the HMIS export.
}
