% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph_destination.R
\name{ph_destination}
\alias{ph_destination}
\title{Permanent Housing Destination HMIS Codes, Labels and Data Frame}
\usage{
ph_destination(input, flag = NULL, ..., .FY = 24)
}
\arguments{
\item{input}{The "Destination Label" as a string; The HMIS CSV Code as an
integer; \code{"ls"} to get the full list of Destination Labels; or \code{"load"} to
get the full PH Destinations data frame.}

\item{flag}{OPTIONAL: For the FY24 and later versions of this function,
adding the flag \code{"category"} will return the category of the input value
(i.e. homeless situation, permanent housing destination, etc.)}

\item{...}{These dots are for future extensions and must be empty.}

\item{.FY}{OPTIONAL: The HMIS data standards fiscal year (defaults to
current).}
}
\value{
Depending on the input: The "Destination" label; The HMIS CSV Code;
the list of PH Destination labels; or the Destination Labels data frame.
}
\description{
If the input is a Destination Label, returns the HMIS CSV Code for that
destination. If the input is an HMIS CSV Code, returns the Destination Label.
If the input is \code{"ls"} returns the list of PH Destination labels. If the
input is \code{"load"} returns the full PH Destinations data frame.
}
