% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_ca_bnl.R
\name{save_ca_bnl}
\alias{save_ca_bnl}
\title{Saves a CE BNL in .rda file format.}
\usage{
save_ca_bnl(
  downloaded_bnl,
  list_date,
  directory_to_save = NULL,
  ...,
  autoclean = FALSE
)
}
\arguments{
\item{downloaded_bnl}{The \code{.xls} or \code{.xlsx} BNL file -or- the folder with the
individual \code{.csv} files of the BNL.}

\item{list_date}{The date of the BNL.}

\item{directory_to_save}{Which folder to save this in. Defaults to the
"CE_BNLs" folder in the "Datasets" folder in the R&E Dropbox.}

\item{...}{These dots are for future extensions and must be empty.}

\item{autoclean}{Automatically filters down the full BNL to include only
certain columns.}
}
\value{
Saves the CE BNL in \code{.rda} format.
}
\description{
NOTE: This requires you to have a RStudio Project open and use the \link{here}
package!
}
