% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_hmis_extract.R
\name{save_hmis_extract}
\alias{save_hmis_extract}
\title{Saves an HMIS CSV/XML Export in .rda file format.}
\usage{
save_hmis_extract(
  extract_path,
  save_in_directory,
  ...,
  full_extract = FALSE,
  .FY = 24
)
}
\arguments{
\item{extract_path}{the path to the folder containing the HMIS CSV/XML Export
.csv files.}

\item{save_in_directory}{the directory in which you want to save the output
file.}

\item{...}{These dots are for future extensions and must be empty.}

\item{full_extract}{Defaults to \code{FALSE}. If set to \code{TRUE} it will include all
14 CSV files from the XML extract instead of the usual 9.}

\item{.FY}{OPTIONAL: the fiscal year of the data standards. Defaults to
current.}

\item{extract_date}{the date the CSV/XML extract was pulled from HMIS:
YYYY-MM-DD.}
}
\value{
saves a file called "hmisData_\strong{extract_date}.rda" with 9 \code{tibble}s:
\strong{client}, \strong{services}, \strong{entry}, \strong{exit},
\strong{project}, \strong{healthanddv},
\strong{incomebenefits}, \strong{organization}, and \strong{disabilities}.
}
\description{
Uses the \link{import_hmis} funtion to read in \code{.csv} files from data exported
from HMIS per the specified date and saves each \code{tibble} to an
"hmisData_\strong{extract_date}.rda" file instead of returning a list of
\code{tibble}s. NOTE: This requires you to have a RStudio Project open and use the
\link{here} package!
}
