% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcut.R
\name{shortcut}
\alias{shortcut}
\title{Quickly Navigate to Certain Folders in the Research-and-Evaluation-Team
Filesystem}
\usage{
shortcut(bookmarkedDirectory, ..., dropbox_username = NULL)
}
\arguments{
\item{bookmarkedDirectory}{as a \code{string}, the abbreviated name of the folder
you want or which your file is in: \code{"echo dropbox"}, \code{"re dropbox"},
\code{"extracts"}, \code{"sdr data"}, \code{"bnl"}, \code{"bnl excel"}, \code{"db inputs"}, \code{"db outputs"}. You can also type \code{"newest extract"} to get the direct link to
the most recent HMIS extract or \code{"newest bnl"} to get the most recent BNL.}

\item{...}{the rest of the path to you file as you would type it in the
\link{here} function.}
}
\value{
Returns the \link{here} function with the first part filled in up to the
parent destination folder (\code{"extracts"}/\code{"bnl"}/etc) with the rest appended
to reach your destination file, so for example you could type
\code{shortcut("bnl", "XXXXXXXX CA BNL.xlsx")} from the Dashboard .rproject
instead of \verb{here("../", "Project Requests", "Small Data Requests",. "data" "XXXXXXXX CA BNL.xlsx")}.
}
\description{
\emph{Shortcut} is a quicker way to use the \link{here} function, customized to
provide quick access to the ECHO Dropbox, RE Team Dropbox, Small Data
Requests (SDR) data folder, HMIS extracts (in Datasets and in SDR/data), CE
BNL (in Datasets and in SDR/data), and Dashboard Input/Output folders.
}
