% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_in_excel.R
\name{show_in_excel}
\alias{show_in_excel}
\title{Open a Data Frame in Excel}
\usage{
show_in_excel(.data)
}
\arguments{
\item{.data}{the data frame to open in Excel.}
}
\value{
Opens \code{.data} in Excel.
}
\description{
Uses MS Excel to open the data frame input as an argument.
}
