% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_project_name_id.R
\name{split_project_name_id}
\alias{split_project_name_id}
\title{Split A single Column Containing Project Name & Project ID into Separate
Columns}
\usage{
split_project_name_id(
  .data,
  original_column,
  new_column_name = "ProjectID",
  ...,
  rename_original_column = TRUE
)
}
\arguments{
\item{.data}{The data frame.}

\item{original_column}{The column containing the combined Project Name and
Project ID to be separated.}

\item{new_column_name}{The new column containing only the Project ID. If not
specified, this defaults to \code{ProjectID}.}

\item{...}{These dots are for future extensions and must be empty.}

\item{rename_original_column}{If set to \code{TRUE}, this will rename the original
column to \code{ProjectName} to match what it is called in the HMIS extract (if
it is not already called that). Defaults to \code{FALSE}.}
}
\value{
In the input data frame, a new column (named \code{ProjectID} by default)
containing only the project ID number, and the old column (renamed
\code{ProjectName} by default) containing only the project name.
}
\description{
This function returns a column named \code{ProjectID} (by default) that removes
the 4-digit Project ID number that is sometimes combined with the Project
Name when working with BusinessObjects reports instead of the HMIS extract.
It is a quick alternative to specifying a \code{\link[dplyr:mutate]{dplyr::mutate()}} by hand.
}
