% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_echo_package.R
\name{update_echo_package}
\alias{update_echo_package}
\title{Update the ECHO Package}
\usage{
update_echo_package()
}
\value{
Updates the ECHO package.
}
\description{
This function updates the ECHO package. If you have already run
\code{library(ECHO)}, you will have to do so again after using this function
because it must unload the ECHO package in order to update it.
}
