% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_color.R
\name{use_color}
\alias{use_color}
\title{Use a Specific Color from a Predefined List for Data Visualization}
\usage{
use_color(specific_color)
}
\arguments{
\item{specific_color}{the color you want to use or \code{"ls"} if you need the
list of colors.}
}
\value{
a hex color value or the list of available colors.
}
\description{
Returns the hex value of the color name that is supplied as an argument or
returns the list of colors that may be used if the supplied argument is
\code{"ls"}.
}
