% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_spid.R
\name{view_spid}
\alias{view_spid}
\title{View Client Information by SPID}
\usage{
view_spid(csvextract, uid)
}
\arguments{
\item{csvextract}{the list returned by the \code{\link[=import_hmis]{import_hmis()}} function.}

\item{uid}{the client's SPID.}
}
\value{
the \code{entry}, \code{client}, and \code{exit} data frames filtered by SPID.
}
\description{
Search for entries, exits, and client demographics by ServicePoint ID number.
}
